riscv: Add support for SPI on Kendryte K210
This enables configs necessary for using SPI. The environment is saved to
the very end of SPI flash. This is unlikely to be overwritten unless the
entire flash is reprogrammed.

This also supplies a default bootcommand. It loads an image and device tree
from the first partition of the MMC. This is a minimal/least effort
bootcmd, so suggestions (especially in the form of patches) are welcome. I
didn't set up distro boot because I think it is unlikely that any
general-purpose linux distros will ever be ported to this board.

Signed-off-by: Sean Anderson <seanga2@gmail.com>
Patch-cc: rchen
Series-to: uboot
Series-cc: jteki
Series-cc: marex
Series-version: 1
Series-process-log: sort
Commit-notes:
Sorry for the late follow-up Jagen.

This patch was previously part of
https://patchwork.ozlabs.org/project/uboot/list/?series=208443
END