reset: Add generic reset driver
This patch adds a generic reset driver. It is designed to be useful when
one has a register in a regmap which contains bits that reset other
devices. I thought this seemed like a very generic use, so here is a
generic driver. The overall structure has been modeled on the syscon-reboot
driver.

Signed-off-by: Sean Anderson <seanga2@gmail.com>
Patch-cc: sjg
Series-changes: 3
- New

Series-changes: 4
- Added basic test
- Fix incorrect usage of regmap_update_bits