dm: Add support for simple-pm-bus
This type of bus is used in Linux to designate buses which have power
domains and/or clocks which need to be enabled before their child devices
can be used. Because power domains are automatically enabled before probing
in U-Boot, we just need to enable any clocks present.

Signed-off-by: Sean Anderson <seanga2@gmail.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Patch-cc: sjg
Series-changes: 3
- New

Series-changes: 4
- Split the bus off into its own driver
- Add test
- Fix line spacing in Kconfig
- Lint

Series-changes: 5
- Reorder includes (simple pm)