spi: dw: Add mem_ops
The dw spi devices on the Kendryte K210 must be operated in a specific
fasion which cannot be achived through multiple writes to via dw_spi_xfer
(as it is currently written). This patch adds an implementation of exec_op,
which gives correct behaviour when reading/writing spi flash.

I would like to be able to modify the existing dw_spi_xfer function such
that it works properly (e.g. with the mmc_spi driver). However, the only
example code I have to work off is Kendryte's sdk (which is written in the
exec_op style), and I do not have access to the datasheet (if anyone does,
I would love to have a look!).

Signed-off-by: Sean Anderson <seanga2@gmail.com>
Patch-cc: sjg
Patch-cc: jteki
Patch-cc: Eugeniy Paltsev <Eugeniy.Paltsev@synopsys.com>
Series-changes: 4
- New