spi: dw: Add mem_ops
The designware ssi device has "broken" chip select behaviour [1], and needs
specific manipulation to use the built-in chip select. The existing fix is
to use an external GPIO for chip select, but typically the K210 has SPI3
directly connected to a flash chip with dedicated pins. This makes it
impossible to use the spi_xfer function to use spi, since the CS is
de-asserted in between calls.  This patch adds an implementation of
exec_op, which gives correct behaviour when reading/writing spi flash.

This patch also rearranges the headers to conform to U-Boot style.

[1] https://lkml.org/lkml/2015/12/23/132

Signed-off-by: Sean Anderson <seanga2@gmail.com>
Tested-by Eugeniy Paltsev <Eugeniy.Paltsev@synopsys.com>
Commit-notes:
This patch was previously part of
https://patchwork.ozlabs.org/project/uboot/list/?series=161576
END
Commit-changes: 2
- Convert debug to log_*

Series-changes: 2
- Add external gpio cs support
- Clean up exec_op
- Limit data transfers to 64k

Series-changes: 3
- Use constant 0x10000 instead of SZ_64K. The latter is not included on
  some platforms and I'm too lazy to figure out what the correct header is.

Commit-changes: 4
- Overcome my sloth

Series-changes: 4
- Rearrange headers in designware_spi.c