patman: Suppress empty changelog entries
Patman outputs a line for every edition of the series in every patch,
regardless of whether any changes were made. This can result in many
redundant lines in patch changelogs, especially when a patch did not exist
before a certain revision. For example, the existing behaviour could result
in a changelog of

Changes in v7: None
Changes in v6: None
Changes in v5:
- Make a change

Changes in v4: None

Changes in v3:
- New

Changes in v2: None

With this patch applied and with --no-empty-changes, the same patch would
look like

(no changes since v5)

Changes in v5:
- Make a change

Changes in v3:
- New

This is entirely aesthetic, but I think it reduces clutter, especially for
patches added later on in a series.

Signed-off-by: Sean Anderson <seanga2@gmail.com>
Series-changes: 2
- Add a note when there are no changes in the current revision

Commit-changes: 2
- Make this the default behaviour, and remove the option

Series-changes: 3
- Fix KeyError when running tests
- Document empty changelog suppression in README

Commit-changes: 3
- Make the newest_version logic more robust (and ugly)
- Fix no changes message sometimes being output before every
  non-newest-revision change
- Fix no changes message being output for revision 1
- Update commit subject

Cover-changes: 3
- Fix some corner cases for no changes messages
- Update commit subjects