spi: dw: Convert calls to debug to log_*
This allows different log levels to be enabled or disabled depending on the
desired level of verbosity. In particular, it allows for general debug
information to be printed while excluding more verbose logging which may
interfere with timing.

log_xxx was chosen over dev_xxx because it shows more information (such as
the calling function) and it can be enabled and disabled at run-time.

Signed-off-by: Sean Anderson <seanga2@gmail.com>
Tested-by Eugeniy Paltsev <Eugeniy.Paltsev@synopsys.com>
Commit-changes: 2
- New

Cover-changes: 2
- Convert debug calls to log_ instead of removing the ones which affect
  timing

Series-changes: 3
- Lower the log level of some messages
- Prefix user-facing logs with SPI@<address>
- Reword error messages as "message (error %d)"