riscv: Don't reserve AI ram in k210 dts
It is no longer necessary to disallow ai ram, since it is enabled by the
sram driver.

Signed-off-by: Sean Anderson <seanga2@gmail.com>
Series-to: u-boot
Series-cc: rchen
Series-cc: bmeng
Series-cc: xypron
Series-version: 2
Cover-letter:
riscv: k210: Enable use of AI ram bank
This ram bank was previously unusable because PLL1 was not started correctly.
This series fixes that bug, and allows U-Boot to relocate into the AI ram. This
provides an extra 2M of space in which to load payloads.
END