spi: dw: Use a mux to access registers
To enter XIP mode, the xip_en signal must be asserted. The exact method of
setting xip_en is integration-specific, but on the K210 (and Baikal-T1) it
is set by a bit in a system configuration register. To handle this, use a
mux to select the state of xip_en before every access to control registers.

Signed-off-by: Sean Anderson <seanga2@gmail.com>
Patch-cc: Eugeniy Paltsev <Eugeniy.Paltsev@synopsys.com>