cmd: Add support for (limited) command substitution
The traditional way to grab the output from a shell script is to use
command substitution. Unfortunately, we don't really have the concept of
pipes in U-Boot (at least not without console recording). Even if we did,
stdout is severely polluted by informational printfs.

Instead of redirecting stdout, instead use a special global variable. This
lets commands set a result without having to modify the function signature
of every command, and everything that calls commands. This is a bit of a
hack, but seemed like the least-invasive of all options.

Currently, the value of cmd_result is printed if it is set. This makes
sense for things like echo, where you can do something like

	=> echo a b c d e
	a b c d e
	=> var=c
	=> echo a $(echo b $var d) e
	a b c d e

but it makes less sense for some other commands

All callers of cmd_process must
	1. Print cmd_result (unless it is being "redirected")
	2. free() cmd_result
	3. set cmd_result to NULL
Calling cmd_process with a non-NULL value in cmd_result is a bug.

Signed-off-by: Sean Anderson <seanga2@gmail.com>