test: hush: Add test for command substitution
This adds a basic test for command substitution. A more advanced test
currently fails, and is left commented-out for now. A basic example of the
difference is that in bash

	$ a=b; echo a=$(echo $a)
	a=b

but in hush

	=> a=b; echo a=$(echo $a)
	a=

This is because command substitution takes place during parsing, before any
of the rest of the command has been executed.

Signed-off-by: Sean Anderson <seanga2@gmail.com>