cli: lil: Handle unbalanced brackets and quotes
Instead of ignoring unbalanced brackets and quotes (behaving as if there
were appropriate balancing brackets and quotes at the end of input),
instead create an error. We use a new error code to make it easier to
implement an interactive shell which prompts for additional input.

Unfortunately, we don't really have a separate parsing and execution step.
So if you do something like

	echo "[echo a; return b]

then 'a' will still be printed, even though the quote is not balanced. This
would likely require a major rewrite to fix.

Signed-off-by: Sean Anderson <seanga2@gmail.com>